requirements.txt
--prefer-binary
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Login==0.6.2
Flask-CORS==4.0.0
python-socketio==5.8.0
Werkzeug==2.3.7
cryptography==39.0.2
gunicorn==20.1.0
python-dotenv==0.19.2
setuptools==68.2.2
supabase==2.3.1
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7